feat(react): add actionList
