+++ packages/react/src/components/pagination/components/Pagination.stories.tsx
@@ -46,26 +46,49 @@ export const WithPaginationDetail = () => {
   const [limit, setLimit] = React.useState(10);
 
   return (
-    <Box display="flex" justifyContent="space-between" alignItems="center">
-      <PaginationDetail page={currentPage} limit={limit} total={50 * limit} length={limit} />
-      <Pagination
-        current={currentPage}
-        total={50}
-        onSelect={select => {
-          setCurrentPage(select);
-          action('select-page')(select);
-        }}
-      />
-      <PaginationFilter
-        selectedItem={limit}
-        handleSelectedItemChange={({ selectedItem }) => {
-          if (selectedItem) {
-            setLimit(selectedItem);
-          }
-        }}
-        items={[10, 20, 30, 40, 50]}
-        placeholder="limit"
-      />
+    <Box display="flex" flexDirection="column" height="80vh" justifyContent="space-between">
+      <Box display="flex" justifyContent="space-between" alignItems="center">
+        <PaginationDetail page={currentPage} limit={limit} total={50 * limit} length={limit} />
+        <Pagination
+          current={currentPage}
+          total={50}
+          onSelect={select => {
+            setCurrentPage(select);
+            action('select-page')(select);
+          }}
+        />
+        <PaginationFilter
+          selectedItem={limit}
+          handleSelectedItemChange={({ selectedItem }) => {
+            if (selectedItem) {
+              setLimit(selectedItem);
+            }
+          }}
+          items={[10, 20, 30, 40, 50]}
+          placeholder="limit"
+        />
+      </Box>
+      <Box display="flex" justifyContent="space-between" alignItems="center">
+        <PaginationDetail page={currentPage} limit={limit} total={50 * limit} length={limit} />
+        <Pagination
+          current={currentPage}
+          total={50}
+          onSelect={select => {
+            setCurrentPage(select);
+            action('select-page')(select);
+          }}
+        />
+        <PaginationFilter
+          selectedItem={limit}
+          handleSelectedItemChange={({ selectedItem }) => {
+            if (selectedItem) {
+              setLimit(selectedItem);
+            }
+          }}
+          items={[10, 20, 30, 40, 50]}
+          placeholder="limit"
+        />
+      </Box>
     </Box>
   );
 };
