+++ packages/react/src/components/pagination/components/PaginationFilter.tsx
@@ -1,7 +1,11 @@
+import { IconChevronDown } from '@aksara-ui/icons';
 import * as React from 'react';
 import Box from '../../../layout/box/components/Box';
 import { Text } from '../../../typography';
-import { InputSelect, InputSelectProps } from '../../form';
+import { Button } from '../../button';
+import ActionList from '../../dropdown/components/DropdownMenuItem/ActionList';
+import { InputSelectProps } from '../../form';
+import { Popover, PopoverContent, PopoverTrigger } from '../../popover';
 
 export interface PaginationFilterProps extends InputSelectProps<number> {
   /** Total limit of pages. */
@@ -25,13 +29,20 @@ const PaginationFilter: React.FC<PaginationFilterProps> = ({
       <Text scale={300} mr={10} {...rest}>
         {label}
       </Text>
-      <InputSelect
-        selectedItem={selectedItem}
-        placeholder={placeholder}
-        width={width}
-        items={items}
-        handleSelectedItemChange={handleSelectedItemChange}
-      />
+      <Popover>
+        <PopoverTrigger>
+          <Button type="button" size="md" icon={IconChevronDown} iconPosition="right">
+            {selectedItem}
+          </Button>
+        </PopoverTrigger>
+        <PopoverContent placement="top">
+          <Box width="100%" maxWidth={300} padding="md">
+            {items.map(item => (
+              <ActionList isActive={item === selectedItem}>{item}</ActionList>
+            ))}
+          </Box>
+        </PopoverContent>
+      </Popover>
     </Box>
   );
 };
